Fix :: addVisibleCount 수정
